+++ api/src/main/kotlin/io/github/teamcheeze/plum/api/core/inventory/GInventoryInteractType.kt
@@ -1,9 +1,0 @@
-package io.github.teamcheeze.plum.api.core.inventory
-
-enum class GInventoryInteractType {
-    CLICK,
-    CLOSE,
-    OPEN,
-    PUT,
-    SELECT
-}