+++ api/src/main/kotlin/io/github/teamcheeze/plum/api/core/inventory/GItem.kt
@@ -1,5 +1,6 @@
 package io.github.teamcheeze.plum.api.core.inventory
 
+import io.github.teamcheeze.plum.api.util.core.Attributable
 import org.bukkit.Material
 import org.bukkit.inventory.ItemFlag
 import org.bukkit.inventory.ItemStack
@@ -8,7 +9,7 @@ import org.bukkit.inventory.ItemStack
  * An ItemStack manager
  * @author dolphin2410
  */
-data class GItem(val item: ItemStack) {
+data class GItem(val item: ItemStack): Attributable<String, Any> {
     companion object {
         /**
          * Create an item with the specified information
@@ -21,13 +22,35 @@ data class GItem(val item: ItemStack) {
          */
         fun create(material: Material, name: String = material.name, enchantment: List<GEnchant> = listOf(), lore: List<String> = listOf(), flag: List<ItemFlag> = listOf()): GItem {
             return GItem(ItemStack(material).apply {
-                itemMeta?.setDisplayName(name)
-                itemMeta?.lore = lore
-                itemMeta?.addItemFlags(*flag.toTypedArray())
+                val newItemMeta = itemMeta
+                newItemMeta?.setDisplayName(name)
+                newItemMeta?.lore = lore
+                newItemMeta?.addItemFlags(*flag.toTypedArray())
                 enchantment.forEach {
-                    itemMeta?.addEnchant(it.enchantment, it.level, it.overrideLevel)
+                    newItemMeta?.addEnchant(it.enchantment, it.level, it.overrideLevel)
                 }
+                itemMeta = newItemMeta
             })
+        }
+    }
+
+    private val internal = HashMap<String, Any>()
+
+    override fun getAttribute(attributeName: String): Any? {
+        return internal[attributeName]
+    }
+
+    override fun setAttribute(key: String, value: Any) {
+        internal[key] = value
+    }
+
+    override fun getAttributes(): Map<String, Any> {
+        return internal
+    }
+
+    override fun setAttributes(vararg pairs: Pair<String, Any>) {
+        pairs.forEach {
+            internal[it.first] = it.second
         }
     }
 }